Tests/LibWeb/Text/expected/css/details-content-computed-style-reflects-state.txt

details-content open content-visibility: visible
details-content closed content-visibility: hidden
